backend/requirements.txt
fastapi
uvicorn
opencv-python
numpy
pillow
python-multipart
httpx
mysql-connector-python
